rename folder to composite
