+++ packages/core/src/composite/composite-modes.js
