+++ packages/core/src/composite/index.js
